fix: setTimeout work
